Apply buffer copy
This change is made in order to have a fair comparison between
exact filter and FIR & multishift filters.

Since FIR and multishift filters use buffer copy previously, and
it yields a computational overhead that is in fact not negligible.
Thus, this buffer copy procedure is implemented in exact DCT
filters.